+++ vite-react/package.json
@@ -7,11 +7,13 @@
     "dev": "vite",
     "build": "vite build",
     "lint": "eslint .",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "start-server": "node simple_api/server.js"
   },
   "dependencies": {
     "@reduxjs/toolkit": "^2.6.1",
     "classnames": "^2.5.1",
+    "express": "^5.1.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-redux": "^9.2.0",
