+++ vite-react/simple_api/server.js
@@ -1,0 +1,27 @@
+const express = require('express');
+const api = require('./api');
+const bodyParser = require('body-parser');
+const port = 3001;
+
+const app = express();
+
+app.use(function (req, res, next) {
+  res.header('Access-Control-Allow-Origin', '*');
+  res.header('Access-Control-Allow-Methods', '*');
+  res.header(
+    'Access-Control-Allow-Headers',
+    'Origin, X-Requested-With, Content-Type, Accept'
+  );
+  next();
+});
+app.use(bodyParser.json());
+app.use('/api', api);
+
+app.listen(port, 'localhost', function (err) {
+  if (err) {
+    console.log(err);
+    return;
+  }
+
+  console.log('Listening at http://localhost:' + port);
+});
